+++ src/ladb_opencutlist/ruby/model/report/report_entry.rb
@@ -71,12 +71,10 @@ module Ladb::OpenCutList
 
   class SheetGoodReportEntry < AbstractReportEntry
 
-    attr_reader :volumic_mass,:std_price, :total_count, :total_area, :total_used_area, :sheets
+    attr_reader :volumic_mass, :std_price, :total_count, :total_area, :total_used_area, :sheets
 
     def initialize(_def)
       super(_def)
-
-      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
 
       @total_count = _def.total_count
       @total_area = _def.total_area == 0 ? nil : DimensionUtils.instance.format_to_readable_area(_def.total_area)
@@ -84,6 +82,7 @@ module Ladb::OpenCutList
 
       @sheets = _def.sheet_defs.values.map { |sheet_def| sheet_def.create_sheet }
 
+      @volumic_mass = @sheets.map { |sheet| sheet.volumic_mass }.select { |volumic_mass| !volumic_mass.nil? }.uniq.join(', ')
       @std_price = @sheets.map { |sheet| sheet.std_price }.select { |std_price| !std_price.nil? }.uniq.join(', ')
 
     end
@@ -92,11 +91,12 @@ module Ladb::OpenCutList
 
   class SheetGoodReportEntrySheet < AbstractReportItem
 
-    attr_reader :std_price, :type, :length, :width, :count, :total_area, :total_used_area
+    attr_reader :volumic_mass, :std_price, :type, :length, :width, :count, :total_area, :total_used_area
 
     def initialize(_def)
       super(_def)
 
+      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
       @std_price = _def.std_price.nil? || _def.std_price[:val] == 0 ? nil : UnitUtils.format_readable(_def.std_price[:val], _def.std_price[:unit], 2, 2)
 
       @type = _def.type
@@ -120,14 +120,13 @@ module Ladb::OpenCutList
     def initialize(_def)
       super(_def)
 
-      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
-
       @total_count = _def.total_count
       @total_length = _def.total_length == 0 ? nil : DimensionUtils.instance.format_to_readable_length(_def.total_length)
       @total_used_length = _def.total_length == 0 ? nil : DimensionUtils.instance.format_to_readable_length(_def.total_used_length)
 
       @bars = _def.bar_defs.values.map { |bar_def| bar_def.create_bar }
 
+      @volumic_mass = @bars.map { |bar| bar.volumic_mass }.select { |volumic_mass| !volumic_mass.nil? }.uniq.join(', ')
       @std_price = @bars.map { |bar| bar.std_price }.select { |std_price| !std_price.nil? }.uniq.join(', ')
 
     end
@@ -136,11 +135,12 @@ module Ladb::OpenCutList
 
   class DimensionalReportEntryBar < AbstractReportItem
 
-    attr_reader :std_price, :type, :length, :count, :total_length, :total_used_length
+    attr_reader :volumic_mass, :std_price, :type, :length, :count, :total_length, :total_used_length
 
     def initialize(_def)
       super(_def)
 
+      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
       @std_price = _def.std_price.nil? || _def.std_price[:val] == 0 ? nil : UnitUtils.format_readable(_def.std_price[:val], _def.std_price[:unit], 2, 2)
 
       @type = _def.type
@@ -163,14 +163,13 @@ module Ladb::OpenCutList
     def initialize(_def)
       super(_def)
 
-      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
-
       @total_count = _def.total_count
       @total_length = _def.total_length == 0 ? nil : DimensionUtils.instance.format_to_readable_length(_def.total_length)
       @total_used_length = _def.total_length == 0 ? nil : DimensionUtils.instance.format_to_readable_length(_def.total_used_length)
 
       @bars = _def.bar_defs.values.map { |bar_def| bar_def.create_bar }
 
+      @volumic_mass = @bars.map { |bar| bar.volumic_mass }.select { |volumic_mass| !volumic_mass.nil? }.uniq.join(', ')
       @std_price = @bars.map { |bar| bar.std_price }.select { |std_price| !std_price.nil? }.uniq.join(', ')
 
     end
@@ -179,11 +178,12 @@ module Ladb::OpenCutList
 
   class EdgeReportEntryBar < AbstractReportItem
 
-    attr_reader :std_price, :type, :length, :count, :total_length, :total_used_length
+    attr_reader :volumic_mass, :std_price, :type, :length, :count, :total_length, :total_used_length
 
     def initialize(_def)
       super(_def)
 
+      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
       @std_price = _def.std_price.nil? || _def.std_price[:val] == 0 ? nil : UnitUtils.format_readable(_def.std_price[:val], _def.std_price[:unit], 2, 2)
 
       @type = _def.type
@@ -246,12 +246,10 @@ module Ladb::OpenCutList
 
   class VeneerReportEntry < AbstractReportEntry
 
-    attr_reader :volumic_mass,:std_price, :total_count, :total_area, :total_used_area, :sheets
+    attr_reader :volumic_mass, :std_price, :total_count, :total_area, :total_used_area, :sheets
 
     def initialize(_def)
       super(_def)
-
-      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
 
       @total_count = _def.total_count
       @total_area = _def.total_area == 0 ? nil : DimensionUtils.instance.format_to_readable_area(_def.total_area)
@@ -259,6 +257,7 @@ module Ladb::OpenCutList
 
       @sheets = _def.sheet_defs.values.map { |sheet_def| sheet_def.create_sheet }
 
+      @volumic_mass = @bars.map { |sheet| sheet.volumic_mass }.select { |volumic_mass| !volumic_mass.nil? }.uniq.join(', ')
       @std_price = @sheets.map { |sheet| sheet.std_price }.select { |std_price| !std_price.nil? }.uniq.join(', ')
 
     end
@@ -267,11 +266,12 @@ module Ladb::OpenCutList
 
   class VeneerReportEntrySheet < AbstractReportItem
 
-    attr_reader :std_price, :type, :length, :width, :count, :total_area, :total_used_area
+    attr_reader :volumic_mass, :std_price, :type, :length, :width, :count, :total_area, :total_used_area
 
     def initialize(_def)
       super(_def)
 
+      @volumic_mass = _def.volumic_mass.nil? || _def.volumic_mass[:val] == 0 ? nil : UnitUtils.format_readable(_def.volumic_mass[:val], _def.volumic_mass[:unit])
       @std_price = _def.std_price.nil? || _def.std_price[:val] == 0 ? nil : UnitUtils.format_readable(_def.std_price[:val], _def.std_price[:unit], 2, 2)
 
       @type = _def.type