+++ src/ladb_opencutlist/ruby/model/report/report_entry_def.rb
@@ -68,13 +68,11 @@ module Ladb::OpenCutList
 
   class SheetGoodReportEntryDef < AbstractReportEntryDef
 
-    attr_accessor :volumic_mass, :total_count, :total_area, :total_used_area
+    attr_accessor :total_count, :total_area, :total_used_area
     attr_reader :sheet_defs
 
     def initialize(cutlist_group)
       super(cutlist_group)
-
-      @volumic_mass = nil
 
       @total_count = 0
       @total_area = 0
@@ -94,12 +92,13 @@ module Ladb::OpenCutList
 
   class SheetGoodReportEntrySheetDef < AbstractReportItemDef
 
-    attr_accessor :std_price, :count, :total_area, :total_used_area
+    attr_accessor :volumic_mass, :std_price, :count, :total_area, :total_used_area
     attr_reader :type, :length, :width
 
     def initialize(cuttingdiagram2d_sheet)
       super()
 
+      @volumic_mass = nil
       @std_price = nil
 
       @type = cuttingdiagram2d_sheet.def.type
@@ -124,13 +123,11 @@ module Ladb::OpenCutList
 
   class DimensionalReportEntryDef < AbstractReportEntryDef
 
-    attr_accessor :volumic_mass, :total_count, :total_length, :total_used_length
+    attr_accessor :total_count, :total_length, :total_used_length
     attr_reader :bar_defs
 
     def initialize(cutlist_group)
       super(cutlist_group)
-
-      @volumic_mass = nil
 
       @total_count = 0
       @total_length = 0
@@ -150,12 +147,13 @@ module Ladb::OpenCutList
 
   class DimensionalReportEntryBarDef < AbstractReportItemDef
 
-    attr_accessor :count, :std_price, :total_length, :total_used_length
+    attr_accessor :volumic_mass, :std_price, :count, :total_length, :total_used_length
     attr_reader :type, :length
 
     def initialize(cuttingdiagram1d_bar)
       super()
 
+      @volumic_mass = nil
       @std_price = nil
 
       @type = cuttingdiagram1d_bar.def.type
@@ -179,13 +177,11 @@ module Ladb::OpenCutList
 
   class EdgeReportEntryDef < AbstractReportEntryDef
 
-    attr_accessor :volumic_mass, :total_count, :total_length, :total_used_length
+    attr_accessor :total_count, :total_length, :total_used_length
     attr_reader :bar_defs
 
     def initialize(cutlist_group)
       super(cutlist_group)
-
-      @volumic_mass = nil
 
       @total_count = 0
       @total_length = 0
@@ -205,12 +201,13 @@ module Ladb::OpenCutList
 
   class EdgeReportEntryBarDef < AbstractReportItemDef
 
-    attr_accessor :count, :std_price, :total_length, :total_used_length
+    attr_accessor :volumic_mass, :std_price, :count, :total_length, :total_used_length
     attr_reader :type, :length
 
     def initialize(cuttingdiagram1d_bar)
       super()
 
+      @volumic_mass = nil
       @std_price = nil
 
       @type = cuttingdiagram1d_bar.def.type
@@ -287,13 +284,11 @@ module Ladb::OpenCutList
 
   class VeneerReportEntryDef < AbstractReportEntryDef
 
-    attr_accessor :volumic_mass, :total_count, :total_area, :total_used_area
+    attr_accessor :total_count, :total_area, :total_used_area
     attr_reader :sheet_defs
 
     def initialize(cutlist_group)
       super(cutlist_group)
-
-      @volumic_mass = nil
 
       @total_count = 0
       @total_area = 0
@@ -313,12 +308,13 @@ module Ladb::OpenCutList
 
   class VeneerReportEntrySheetDef < AbstractReportItemDef
 
-    attr_accessor :std_price, :count, :total_area, :total_used_area
+    attr_accessor :volumic_mass, :std_price, :count, :total_area, :total_used_area
     attr_reader :type, :length, :width
 
     def initialize(cuttingdiagram2d_sheet)
       super()
 
+      @volumic_mass = nil
       @std_price = nil
 
       @type = cuttingdiagram2d_sheet.def.type