+++ src/ladb_opencutlist/ruby/worker/materials/materials_list_worker.rb
@@ -82,7 +82,7 @@ module Ladb::OpenCutList
                       :std_sizes => material_attributes.std_sizes,
                       :grained => material_attributes.grained,
                       :edge_decremented => material_attributes.edge_decremented,
-                      :volumic_mass => material_attributes.volumic_mass,
+                      :std_volumic_masses => material_attributes.std_volumic_masses,
                       :std_prices => material_attributes.std_prices,
                   }
               }